requirements.txt

fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
boto3==1.28.0
google-cloud-vision==3.4.0
langchain==0.0.200
langchain-openai==0.1.0
